requirements.txt
tensorflow>=2.8.0
opencv-python>=4.5.5.64
numpy>=1.21.5
mediapipe>=0.8.9.1
scikit-learn>=1.0.2
Flask>=2.0.1
pymongo>=4.0.1
python-dotenv>=0.19.2
twilio>=7.7.1
geopy>=2.2.0
flask-mail>=0.9.1
Werkzeug>=2.0.2
requests>=2.27.1
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.1
gdown>=5.1.0
tqdm>=4.66.1 